requirements.txt
awsebcli==3.15.1
blessed==1.15.0
botocore==1.12.149
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
dnspython==1.16.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mongoengine==0.17.0
nltk==3.4.1
numpy==1.16.3
passlib==1.7.1
pathspec==0.5.9
pdfkit==0.6.1
pdfminer.six==20181108
pycryptodome==3.8.1
pymongo==3.8.0
pyparsing==2.4.0
python-aiml==0.9.1
python-dateutil==2.8.0
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sortedcontainers==2.1.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
boto3==1.9.157
razorpay==1.2.0